Simulates the cohort trial.

Simulates the cohort trial.

simulate_trial(
  n_int = 50,
  n_fin = 100,
  cohorts_start = 1,
  rr_comb,
  rr_mono,
  rr_back,
  rr_plac,
  rr_transform,
  random_type = NULL,
  trial_struc = "all_plac",
  random = FALSE,
  prob_comb_rr = NULL,
  prob_mono_rr = NULL,
  prob_back_rr = NULL,
  prob_plac_rr = NULL,
  prob_rr_transform = prob_rr_transform,
  stage_data = TRUE,
  cohort_random = NULL,
  cohorts_max = 4,
  sr_drugs_pos = 1,
  sr_pats = cohorts_max * (n_fin + 3 * cohorts_max),
  sr_first_pos = FALSE,
  target_rr = c(0, 0, 1),
  cohort_offset = 0,
  sharing_type = "all",
  safety_prob = 0,
  ...
)

Arguments

n_int
Sample size per cohort to conduct interim analysis

n_fin
Sample size per cohort at final

cohorts_start
Number of cohorts to start the platform with

rr_comb
Response rates of combination therapies

rr_mono
Response rate of mono therapies

rr_back
Response rates of backbone arms

rr_plac
Response rate of the placebo

rr_transform
Function transforming all the above response rates to a vector of four probabilities for the multinomial simulation First element is probability of both failures. Second element is probability of biomarker success and histology failure. Third element is probability of biomarker failure and histology success. Fourth element is probability of both success.

random_type
How should the response rates be drawn randomly? Options are:

"absolute": Specify absolute response rates that will be drawn with a certain probability

"risk_difference": Specify absolute response rates for placebo which will be drawn randomly, plus specify vectors for absolute treatment effects of mono therapies over placebo and for combo over the mono therapies.

"risk_ratio": Specify absolute response rates for placebo which will be drawn randomly, plus specify vectors for relative treatment effects of mono therapies over placebo and for combo over the mono therapies.

"odds_ratios": Specify response rate for placebo, specify odds-ratios for mono therapies (via rr_back and rr_mono) and respective probabilities. On top, specify interaction for the combination therapy via rr_comb with prob_rr_comb. Set: odds_combo = odds_plac * or_mono1 * or_mono2 * rr_comb. If rr_comb > 1 -> synergistic, if rr_comb = 1 -> additive. If rr_comb < 1 -> antagonistic. Default is "NULL".

trial_struc
Trial Structure: "all_plac" = all cohorts have placebo arm

"no_plac" = no cohort has placebo arm

"stop_post_mono" = all cohorts start with placebo arm, but after first mono has been declared successful, newly enrolled cohorts have no more placebo

"stop_post_back" = all cohorts start with placebo arm, but after first backbone has been declared successful, newly enrolled cohorts have no more placebo

random
Should the response rates of the arms be randomly drawn from rr_exp? Default is FALSE.

prob_comb_rr
If random == TRUE, what are the probabilities with which the elements of rr_comb should be drawn?

prob_mono_rr
If random == TRUE, what are the probabilities with which the elements of rr_mono should be drawn?

prob_back_rr
If random == TRUE, what are the probabilities with which the elements of rr_back should be drawn?

prob_plac_rr
If random == TRUE, what are the probabilities with which the elements of rr_plac should be drawn?

prob_rr_transform
If random == TRUE, what are the probabilities with which the elements of rr_transform should be drawn?

stage_data
Should individual stage data be passed along? Default is TRUE

cohort_random
If not NULL, indicates that new arms/cohorts should be randomly started. For every patient, there is a cohort_random probability that a new cohort will be started.

cohorts_max
Maximum number of cohorts that are allowed to be added throughout the trial

sr_drugs_pos
Stopping rule for successful experimental arms; Default = 1

sr_pats
Stopping rule for total number of patients; Default = cohorts_max * n_fin + error term based on randomization

sr_first_pos
Stopping rule for first successful cohort; if TRUE, after first cohort was found to be successful, no further cohorts will be included but cohorts will finish evaluating, unless other stopping rules reached prior. Default is FALSE.

target_rr
What is target to declare a combo a positive? Vector of length 3 giving 1) the threshold by which the combo needs to be better than the monos and 2) the threhsold by which the monos need to be better than the placebo. The third element of the vector specifies the relation, choices are 1=="risk-difference", 2=="risk-ratio" and 3=="odds-ratio". By default: c(0,0, "risk-difference").

cohort_offset
Minimum number of patients between adding new cohorts

sharing_type
Which backbone and placebo data should be used for arm comparisons; Default is "all". Another option is "concurrent" or "dynamic" or "cohort".

safety_prob
Probability for a safety stop after every patient

...
Further arguments to be passed to decision function, such as decision making criteria

Value

List containing: Responses and patients on experimental and control arm, total treatment successes and failures and final p-value

Examples

random <- TRUE

rr_comb <- c(0.25, 0.35, 0.4)
prob_comb_rr <- c(0.4, 0.4, 0.2)
rr_mono <- c(0.15, 0.20, 0.25)
prob_mono_rr <- c(0.2, 0.4, 0.4)
rr_back <- c(0.20, 0.25, 0.30)
prob_back_rr <- c(0.3, 0.4, 0.3)
rr_plac <- c(0.10, 0.12, 0.14)
prob_plac_rr <- c(0.25, 0.5, 0.25)

rr_transform <- list(
  function(x) {return(c(0.75*(1 - x), (1-0.75)*(1-x), (1-0.75)*x, 0.75*x))},
  function(x) {return(c(0.85*(1 - x), (1-0.85)*(1-x), (1-0.85)*x, 0.85*x))}
)
prob_rr_transform <- c(0.5, 0.5)

cohorts_max <- 4
trial_struc <- "stop_post_back"
safety_prob <- 0
sharing_type <- "concurrent"
sr_drugs_pos <- 4
sr_first_pos <- TRUE
n_int <- 50
n_fin <- 100
stage_data <- TRUE
cohort_random <- 0.05
target_rr <- c(0,0,1)
cohort_offset <- 5
random_type <- "risk_difference"

# Vergleich Combo vs Mono
Bayes_Sup1 <- matrix(nrow = 3, ncol = 3)
Bayes_Sup1[1,] <- c(0.00, 0.90, 1.00)
Bayes_Sup1[2,] <- c(0.05, 0.65, 1.00)
Bayes_Sup1[3,] <- c(0.10, 0.50, 1.00)
# Vergleich Combo vs Backbone
Bayes_Sup2 <- matrix(nrow = 3, ncol = 3)
Bayes_Sup2[1,] <- c(0.05, 0.80, 1.00)
Bayes_Sup2[2,] <- c(NA, NA, NA)
Bayes_Sup2[3,] <- c(NA, NA, NA)
# Vergleich Mono vs Placebo
Bayes_Sup3 <- matrix(nrow = 3, ncol = 3)
Bayes_Sup3[1,] <- c(0.00, 0.90, 1.00)
Bayes_Sup3[2,] <- c(0.05, 0.65, 1.00)
Bayes_Sup3[3,] <- c(NA, NA, NA)
Bayes_Sup4 <- matrix(nrow = 3, ncol = 3)
Bayes_Sup4[1,] <- c(0.00, 0.90, 1.00)
Bayes_Sup4[2,] <- c(0.05, 0.65, 1.00)
Bayes_Sup4[3,] <- c(NA, NA, NA)
Bayes_Sup <- list(list(Bayes_Sup1, Bayes_Sup2, Bayes_Sup3, Bayes_Sup4),
             list(Bayes_Sup1, Bayes_Sup2, Bayes_Sup3, Bayes_Sup4))

simulate_trial(
n_int = n_int, n_fin = n_fin, trial_struc = trial_struc, random_type = random_type,
rr_comb = rr_comb, rr_mono = rr_mono, rr_back = rr_back, rr_plac = rr_plac,
rr_transform = rr_transform, random = random, prob_comb_rr = prob_comb_rr,
prob_mono_rr = prob_mono_rr, prob_back_rr = prob_back_rr, prob_plac_rr = prob_plac_rr,
stage_data = stage_data, cohort_random = cohort_random, cohorts_max = cohorts_max,
sr_drugs_pos = sr_drugs_pos, target_rr = target_rr, sharing_type = sharing_type,
safety_prob = safety_prob, Bayes_Sup = Bayes_Sup, prob_rr_transform = prob_rr_transform,
cohort_offset = cohort_offset, sr_first_pos = sr_first_pos
)

#> $Trial_Overview
#> $Trial_Overview$Decision
#>      Cohort1    Cohort2    Cohort3    Cohort4   
#> [1,] "CONTINUE" "CONTINUE" "CONTINUE" "CONTINUE"
#> [2,] "STOP_N"   "GO_SUP"   "GO_SUP"   "STOP_N"  
#> 
#> $Trial_Overview$RR_Comb
#> Cohort1 Cohort2 Cohort3 Cohort4 
#>    0.87    0.99    0.92    0.87 
#> 
#> $Trial_Overview$RR_Mono
#> Cohort1 Cohort2 Cohort3 Cohort4 
#>    0.32    0.39    0.32    0.32 
#> 
#> $Trial_Overview$RR_Back
#> Cohort1 Cohort2 Cohort3 Cohort4 
#>    0.32    0.39    0.37    0.32 
#> 
#> $Trial_Overview$RR_Plac
#> Cohort1 Cohort2 Cohort3 Cohort4 
#>    0.12    0.14    0.12    0.12 
#> 
#> $Trial_Overview$RR_Target
#> [1] 0 0 1
#> 
#> $Trial_Overview$N_Cohorts
#> [1] 4
#> 
#> $Trial_Overview$N_Cohorts_First_Suc
#> [1] 4
#> 
#> $Trial_Overview$Total_N_Vector
#>  [1]   4   8  12  16  20  24  28  32  36  40  44  48  52  56  60  64  76 100 140
#> [20] 180 220 244 268 292 316 340 364 388 400 404 408
#> 
#> $Trial_Overview$Final_N_Cohort
#>          Cohort1 Cohort2 Cohort3 Cohort4
#> Combo         33      38      38      37
#> Mono          33      38      38      37
#> Backbone      21      12      12      13
#> Placebo       21      12      12      13
#> 
#> $Trial_Overview$Total_N
#> [1] 408
#> 
#> $Trial_Overview$Total_N_First_Suc
#> [1] 388
#> 
#> $Trial_Overview$Perc_N_Sup_Plac_Th
#> [1] 0.8578431
#> 
#> $Trial_Overview$Perc_N_Sup_Plac_Real
#> [1] 0.8578431
#> 
#> $Trial_Overview$Total_N_Comb
#> [1] 146
#> 
#> $Trial_Overview$Total_N_Mono
#> [1] 146
#> 
#> $Trial_Overview$Total_N_Back
#> [1] 58
#> 
#> $Trial_Overview$Total_N_Plac
#> [1] 58
#> 
#> $Trial_Overview$Total_N_Plac_First_Suc
#> [1] 54
#> 
#> $Trial_Overview$Total_N_Plac_Pool
#> [1] 408
#> 
#> $Trial_Overview$Successes_Hist
#> [1] 211
#> 
#> $Trial_Overview$Successes_Hist_Comb
#> [1] 139
#> 
#> $Trial_Overview$Successes_Hist_Mono
#> [1] 45
#> 
#> $Trial_Overview$Successes_Hist_Back
#> [1] 20
#> 
#> $Trial_Overview$Successes_Hist_Plac
#> [1] 7
#> 
#> $Trial_Overview$Successes_Bio
#> [1] 202
#> 
#> $Trial_Overview$Successes_Bio_Comb
#> [1] 107
#> 
#> $Trial_Overview$Successes_Bio_Mono
#> [1] 56
#> 
#> $Trial_Overview$Successes_Bio_Back
#> [1] 24
#> 
#> $Trial_Overview$Successes_Bio_Plac
#> [1] 15
#> 
#> $Trial_Overview$TP
#> [1] 2
#> 
#> $Trial_Overview$FP
#> [1] 0
#> 
#> $Trial_Overview$TN
#> [1] 0
#> 
#> $Trial_Overview$FN
#> [1] 2
#> 
#> $Trial_Overview$any_P
#> [1] 1
#> 
#> 
#> $Stage_Data
#> $Stage_Data$Cohort1
#> $Stage_Data$Cohort1$decision
#> [1] "CONTINUE" "STOP_N"  
#> 
#> $Stage_Data$Cohort1$alloc_ratio
#> [1] 4 4 1 1
#> 
#> $Stage_Data$Cohort1$n_thresh
#> [1] Inf Inf
#> 
#> $Stage_Data$Cohort1$start_n
#> [1] 0
#> 
#> $Stage_Data$Cohort1$Comb
#> $Stage_Data$Cohort1$Comb$rr
#> [1] 0.87
#> 
#> $Stage_Data$Cohort1$Comb$resp_bio
#>  [1]  1  1  1  0  1  1  0  1  0  1  1  1  1  1  1  1  2  3  2  4  4 NA NA NA NA
#> [26] NA NA NA NA NA NA
#> 
#> $Stage_Data$Cohort1$Comb$resp_hist
#>  [1]  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  3  1  4  4 NA NA NA NA
#> [26] NA NA NA NA NA NA
#> 
#> $Stage_Data$Cohort1$Comb$n
#>  [1]  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  3  4  4  4 NA NA NA NA
#> [26] NA NA NA NA NA NA
#> 
#> 
#> $Stage_Data$Cohort1$Mono
#> $Stage_Data$Cohort1$Mono$rr
#> [1] 0.32
#> 
#> $Stage_Data$Cohort1$Mono$resp_bio
#>  [1]  0  1  0  1  0  1  0  1  0  0  1  0  0  0  0  0  1  1  1  0  1 NA NA NA NA
#> [26] NA NA NA NA NA NA
#> 
#> $Stage_Data$Cohort1$Mono$resp_hist
#>  [1]  0  1  0  1  0  1  0  0  0  0  1  0  0  0  0  0  0  1  1  0  0 NA NA NA NA
#> [26] NA NA NA NA NA NA
#> 
#> $Stage_Data$Cohort1$Mono$n
#>  [1]  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  3  4  4  4 NA NA NA NA
#> [26] NA NA NA NA NA NA
#> 
#> 
#> $Stage_Data$Cohort1$Back
#> $Stage_Data$Cohort1$Back$rr
#> [1] 0.32
#> 
#> $Stage_Data$Cohort1$Back$resp_bio
#>  [1]  1  0  1  0  0  1  0  0  0  1  1  1  0  1  0  0  1  0  0  0  1 NA NA NA NA
#> [26] NA NA NA NA NA NA
#> 
#> $Stage_Data$Cohort1$Back$resp_hist
#>  [1]  1  0  1  0  0  1  0  0  0  1  0  1  0  0  0  0  1  0  0  1  1 NA NA NA NA
#> [26] NA NA NA NA NA NA
#> 
#> $Stage_Data$Cohort1$Back$n
#>  [1]  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 NA NA NA NA
#> [26] NA NA NA NA NA NA
#> 
#> 
#> $Stage_Data$Cohort1$Plac
#> $Stage_Data$Cohort1$Plac$rr
#> [1] 0.12
#> 
#> $Stage_Data$Cohort1$Plac$resp_bio
#>  [1]  0  1  0  0  0  0  0  1  0  0  0  1  0  0  0  0  0  0  0  0  0 NA NA NA NA
#> [26] NA NA NA NA NA NA
#> 
#> $Stage_Data$Cohort1$Plac$resp_hist
#>  [1]  0  0  1  0  0  0  0  1  0  0  0  0  0  0  0  0  0  0  0  0  0 NA NA NA NA
#> [26] NA NA NA NA NA NA
#> 
#> $Stage_Data$Cohort1$Plac$n
#>  [1]  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 NA NA NA NA
#> [26] NA NA NA NA NA NA
#> 
#> 
#> $Stage_Data$Cohort1$prob_sup1
#> [1] 0.9774979 0.9111802 0.8015453 0.8921611 1.0000000 0.9999981 0.8590000
#> [8] 0.9929958
#> 
#> $Stage_Data$Cohort1$prob_sup2
#> [1] 0.9588310        NA 0.7107846 0.8307261 1.0000000        NA 0.6766173
#> [8] 0.9744772
#> 
#> $Stage_Data$Cohort1$prob_sup3
#> [1] 0.9290879        NA        NA        NA 1.0000000        NA        NA
#> [8]        NA
#> 
#> $Stage_Data$Cohort1$sup_interim_list
#> $Stage_Data$Cohort1$sup_interim_list[[1]]
#>      [,1] [,2]  [,3]  [,4]
#> [1,] TRUE TRUE FALSE FALSE
#> [2,] TRUE   NA  TRUE  TRUE
#> [3,] TRUE   NA    NA    NA
#> 
#> 
#> $Stage_Data$Cohort1$fut_interim_list
#> list()
#> 
#> $Stage_Data$Cohort1$prom_interim_list
#> $Stage_Data$Cohort1$prom_interim_list[[1]]
#>       [,1]  [,2]  [,3]  [,4]
#> [1,] FALSE FALSE FALSE FALSE
#> [2,] FALSE    NA FALSE FALSE
#> [3,] FALSE    NA    NA    NA
#> 
#> 
#> $Stage_Data$Cohort1$sup_interim
#> [1] FALSE
#> 
#> $Stage_Data$Cohort1$fut_interim
#> [1] FALSE
#> 
#> $Stage_Data$Cohort1$prom_interim
#> [1] FALSE
#> 
#> $Stage_Data$Cohort1$interim_n
#> [1] 52
#> 
#> $Stage_Data$Cohort1$interim_n_cohort
#> [1] 52
#> 
#> $Stage_Data$Cohort1$sup_final_list
#> $Stage_Data$Cohort1$sup_final_list[[1]]
#>      [,1] [,2]  [,3] [,4]
#> [1,] TRUE TRUE FALSE TRUE
#> [2,] TRUE   NA  TRUE TRUE
#> [3,] TRUE   NA    NA   NA
#> 
#> 
#> $Stage_Data$Cohort1$fut_final_list
#> list()
#> 
#> $Stage_Data$Cohort1$prom_final_list
#> $Stage_Data$Cohort1$prom_final_list[[1]]
#>       [,1]  [,2]  [,3]  [,4]
#> [1,] FALSE FALSE FALSE FALSE
#> [2,] FALSE    NA FALSE FALSE
#> [3,] FALSE    NA    NA    NA
#> 
#> 
#> $Stage_Data$Cohort1$sup_final
#> [1] FALSE
#> 
#> $Stage_Data$Cohort1$fut_final
#> [1] FALSE
#> 
#> $Stage_Data$Cohort1$prom_final
#> [1] FALSE
#> 
#> $Stage_Data$Cohort1$final_n
#> [1] 220
#> 
#> $Stage_Data$Cohort1$final_n_cohort
#> [1] 108
#> 
#> 
#> $Stage_Data$Cohort2
#> $Stage_Data$Cohort2$decision
#> [1] "CONTINUE" "GO_SUP"  
#> 
#> $Stage_Data$Cohort2$alloc_ratio
#> [1] 3 3 1 1
#> 
#> $Stage_Data$Cohort2$n_thresh
#> [1] Inf Inf
#> 
#> $Stage_Data$Cohort2$start_n
#> [1] 64
#> 
#> $Stage_Data$Cohort2$Comb
#> $Stage_Data$Cohort2$Comb$rr
#> [1] 0.99
#> 
#> $Stage_Data$Cohort2$Comb$resp_bio
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  1  2  4  2  3  2  2  3  2
#> [26]  1  1  2 NA NA NA
#> 
#> $Stage_Data$Cohort2$Comb$resp_hist
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  2  3  4  4  4  3  3  3  3
#> [26]  3  3  3 NA NA NA
#> 
#> $Stage_Data$Cohort2$Comb$n
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  2  3  4  4  4  3  3  3  3
#> [26]  3  3  3 NA NA NA
#> 
#> 
#> $Stage_Data$Cohort2$Mono
#> $Stage_Data$Cohort2$Mono$rr
#> [1] 0.39
#> 
#> $Stage_Data$Cohort2$Mono$resp_bio
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  1  2  1  1  1  1  1  1  1
#> [26]  1  1  2 NA NA NA
#> 
#> $Stage_Data$Cohort2$Mono$resp_hist
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  2  2  2  2  2  0  0  0  2
#> [26]  1  0  2 NA NA NA
#> 
#> $Stage_Data$Cohort2$Mono$n
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  2  3  4  4  4  3  3  3  3
#> [26]  3  3  3 NA NA NA
#> 
#> 
#> $Stage_Data$Cohort2$Back
#> $Stage_Data$Cohort2$Back$rr
#> [1] 0.39
#> 
#> $Stage_Data$Cohort2$Back$resp_bio
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  0  1  0  1  0  0  0  0  0
#> [26]  1  1  0 NA NA NA
#> 
#> $Stage_Data$Cohort2$Back$resp_hist
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  0  0  0  0  0  0  0  0  0
#> [26]  1  1  0 NA NA NA
#> 
#> $Stage_Data$Cohort2$Back$n
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  1  1  1  1  1  1  1  1  1
#> [26]  1  1  1 NA NA NA
#> 
#> 
#> $Stage_Data$Cohort2$Plac
#> $Stage_Data$Cohort2$Plac$rr
#> [1] 0.14
#> 
#> $Stage_Data$Cohort2$Plac$resp_bio
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  0  0  0  0  1  1  1  1  1
#> [26]  0  0  1 NA NA NA
#> 
#> $Stage_Data$Cohort2$Plac$resp_hist
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  0  0  0  0  0  0  1  1  0
#> [26]  0  0  1 NA NA NA
#> 
#> $Stage_Data$Cohort2$Plac$n
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  1  1  1  1  1  1  1  1  1
#> [26]  1  1  1 NA NA NA
#> 
#> 
#> $Stage_Data$Cohort2$prob_sup1
#> [1] 0.9872783 0.8990214 0.6317154 0.8362617 1.0000000 1.0000000 0.9958973
#> [8] 0.9921236
#> 
#> $Stage_Data$Cohort2$prob_sup2
#> [1] 0.9720277        NA 0.4947152 0.7396267 1.0000000        NA 0.9839466
#> [8] 0.9715571
#> 
#> $Stage_Data$Cohort2$prob_sup3
#> [1] 0.9438354        NA        NA        NA 1.0000000        NA        NA
#> [8]        NA
#> 
#> $Stage_Data$Cohort2$sup_interim_list
#> $Stage_Data$Cohort2$sup_interim_list[[1]]
#>      [,1] [,2]  [,3]  [,4]
#> [1,] TRUE TRUE FALSE FALSE
#> [2,] TRUE   NA FALSE  TRUE
#> [3,] TRUE   NA    NA    NA
#> 
#> 
#> $Stage_Data$Cohort2$fut_interim_list
#> list()
#> 
#> $Stage_Data$Cohort2$prom_interim_list
#> $Stage_Data$Cohort2$prom_interim_list[[1]]
#>       [,1]  [,2]  [,3]  [,4]
#> [1,] FALSE FALSE FALSE FALSE
#> [2,] FALSE    NA FALSE FALSE
#> [3,] FALSE    NA    NA    NA
#> 
#> 
#> $Stage_Data$Cohort2$sup_interim
#> [1] FALSE
#> 
#> $Stage_Data$Cohort2$fut_interim
#> [1] FALSE
#> 
#> $Stage_Data$Cohort2$prom_interim
#> [1] FALSE
#> 
#> $Stage_Data$Cohort2$interim_n
#> [1] 244
#> 
#> $Stage_Data$Cohort2$interim_n_cohort
#> [1] 52
#> 
#> $Stage_Data$Cohort2$sup_final_list
#> $Stage_Data$Cohort2$sup_final_list[[1]]
#>      [,1] [,2] [,3] [,4]
#> [1,] TRUE TRUE TRUE TRUE
#> [2,] TRUE   NA TRUE TRUE
#> [3,] TRUE   NA   NA   NA
#> 
#> 
#> $Stage_Data$Cohort2$fut_final_list
#> list()
#> 
#> $Stage_Data$Cohort2$prom_final_list
#> $Stage_Data$Cohort2$prom_final_list[[1]]
#>       [,1]  [,2]  [,3]  [,4]
#> [1,] FALSE FALSE FALSE FALSE
#> [2,] FALSE    NA FALSE FALSE
#> [3,] FALSE    NA    NA    NA
#> 
#> 
#> $Stage_Data$Cohort2$sup_final
#> [1] TRUE
#> 
#> $Stage_Data$Cohort2$fut_final
#> [1] FALSE
#> 
#> $Stage_Data$Cohort2$prom_final
#> [1] FALSE
#> 
#> $Stage_Data$Cohort2$final_n
#> [1] 388
#> 
#> $Stage_Data$Cohort2$final_n_cohort
#> [1] 100
#> 
#> 
#> $Stage_Data$Cohort3
#> $Stage_Data$Cohort3$decision
#> [1] "CONTINUE" "GO_SUP"  
#> 
#> $Stage_Data$Cohort3$alloc_ratio
#> [1] 2 2 1 1
#> 
#> $Stage_Data$Cohort3$n_thresh
#> [1] Inf Inf
#> 
#> $Stage_Data$Cohort3$start_n
#> [1] 76
#> 
#> $Stage_Data$Cohort3$Comb
#> $Stage_Data$Cohort3$Comb$rr
#> [1] 0.92
#> 
#> $Stage_Data$Cohort3$Comb$resp_bio
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  2  3  4  1  3  3  3  2
#> [26]  2  1  2  2 NA NA
#> 
#> $Stage_Data$Cohort3$Comb$resp_hist
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  3  4  4  4  3  3  3  3
#> [26]  2  3  3  2 NA NA
#> 
#> $Stage_Data$Cohort3$Comb$n
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  3  4  4  4  3  3  3  3
#> [26]  3  3  3  2 NA NA
#> 
#> 
#> $Stage_Data$Cohort3$Mono
#> $Stage_Data$Cohort3$Mono$rr
#> [1] 0.32
#> 
#> $Stage_Data$Cohort3$Mono$resp_bio
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  1  2  1  4  1  1  1  2
#> [26]  1  3  0  1 NA NA
#> 
#> $Stage_Data$Cohort3$Mono$resp_hist
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  1  2  1  3  1  1  1  1
#> [26]  1  3  0  1 NA NA
#> 
#> $Stage_Data$Cohort3$Mono$n
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  3  4  4  4  3  3  3  3
#> [26]  3  3  3  2 NA NA
#> 
#> 
#> $Stage_Data$Cohort3$Back
#> $Stage_Data$Cohort3$Back$rr
#> [1] 0.37
#> 
#> $Stage_Data$Cohort3$Back$resp_bio
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  0  1  1  1  1  0  1  0
#> [26]  0  0  1  0 NA NA
#> 
#> $Stage_Data$Cohort3$Back$resp_hist
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  0  1  1  1  1  0  0  0
#> [26]  0  0  1  0 NA NA
#> 
#> $Stage_Data$Cohort3$Back$n
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  1  1  1  1  1  1  1  1
#> [26]  1  1  1  1 NA NA
#> 
#> 
#> $Stage_Data$Cohort3$Plac
#> $Stage_Data$Cohort3$Plac$rr
#> [1] 0.12
#> 
#> $Stage_Data$Cohort3$Plac$resp_bio
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  0  0  0  1  0  0  0  0
#> [26]  1  0  0  0 NA NA
#> 
#> $Stage_Data$Cohort3$Plac$resp_hist
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  0  0  0  1  0  0  0  0
#> [26]  1  0  0  0 NA NA
#> 
#> $Stage_Data$Cohort3$Plac$n
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  1  1  1  1  1  1  1  1
#> [26]  1  1  1  1 NA NA
#> 
#> 
#> $Stage_Data$Cohort3$prob_sup1
#> [1] 0.9724801 0.9857567 0.8270234 0.6259328 1.0000000 1.0000000 0.9979407
#> [8] 0.9854234
#> 
#> $Stage_Data$Cohort3$prob_sup2
#> [1] 0.9426057        NA 0.7266107 0.4887899 0.9999999        NA 0.9912865
#> [8] 0.9515428
#> 
#> $Stage_Data$Cohort3$prob_sup3
#> [1] 0.8913932        NA        NA        NA 0.9999989        NA        NA
#> [8]        NA
#> 
#> $Stage_Data$Cohort3$sup_interim_list
#> $Stage_Data$Cohort3$sup_interim_list[[1]]
#>      [,1] [,2]  [,3]  [,4]
#> [1,] TRUE TRUE FALSE FALSE
#> [2,] TRUE   NA  TRUE FALSE
#> [3,] TRUE   NA    NA    NA
#> 
#> 
#> $Stage_Data$Cohort3$fut_interim_list
#> list()
#> 
#> $Stage_Data$Cohort3$prom_interim_list
#> $Stage_Data$Cohort3$prom_interim_list[[1]]
#>       [,1]  [,2]  [,3]  [,4]
#> [1,] FALSE FALSE FALSE FALSE
#> [2,] FALSE    NA FALSE FALSE
#> [3,] FALSE    NA    NA    NA
#> 
#> 
#> $Stage_Data$Cohort3$sup_interim
#> [1] FALSE
#> 
#> $Stage_Data$Cohort3$fut_interim
#> [1] FALSE
#> 
#> $Stage_Data$Cohort3$prom_interim
#> [1] FALSE
#> 
#> $Stage_Data$Cohort3$interim_n
#> [1] 268
#> 
#> $Stage_Data$Cohort3$interim_n_cohort
#> [1] 54
#> 
#> $Stage_Data$Cohort3$sup_final_list
#> $Stage_Data$Cohort3$sup_final_list[[1]]
#>      [,1] [,2] [,3] [,4]
#> [1,] TRUE TRUE TRUE TRUE
#> [2,] TRUE   NA TRUE TRUE
#> [3,] TRUE   NA   NA   NA
#> 
#> 
#> $Stage_Data$Cohort3$fut_final_list
#> list()
#> 
#> $Stage_Data$Cohort3$prom_final_list
#> $Stage_Data$Cohort3$prom_final_list[[1]]
#>       [,1]  [,2]  [,3]  [,4]
#> [1,] FALSE FALSE FALSE FALSE
#> [2,] FALSE    NA FALSE FALSE
#> [3,] FALSE    NA    NA    NA
#> 
#> 
#> $Stage_Data$Cohort3$sup_final
#> [1] TRUE
#> 
#> $Stage_Data$Cohort3$fut_final
#> [1] FALSE
#> 
#> $Stage_Data$Cohort3$prom_final
#> [1] FALSE
#> 
#> $Stage_Data$Cohort3$final_n
#> [1] 400
#> 
#> $Stage_Data$Cohort3$final_n_cohort
#> [1] 100
#> 
#> 
#> $Stage_Data$Cohort4
#> $Stage_Data$Cohort4$decision
#> [1] "CONTINUE" "STOP_N"  
#> 
#> $Stage_Data$Cohort4$alloc_ratio
#> [1] 1 1 1 1
#> 
#> $Stage_Data$Cohort4$n_thresh
#> [1] Inf Inf
#> 
#> $Stage_Data$Cohort4$start_n
#> [1] 100
#> 
#> $Stage_Data$Cohort4$Comb
#> $Stage_Data$Cohort4$Comb$rr
#> [1] 0.87
#> 
#> $Stage_Data$Cohort4$Comb$resp_bio
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  4  3  3  3  1  3  2
#> [26]  2  2  1  1  1  0
#> 
#> $Stage_Data$Cohort4$Comb$resp_hist
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  4  4  4  3  2  3  3
#> [26]  2  2  3  2  1  1
#> 
#> $Stage_Data$Cohort4$Comb$n
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  4  4  4  3  3  3  3
#> [26]  3  3  3  2  1  1
#> 
#> 
#> $Stage_Data$Cohort4$Mono
#> $Stage_Data$Cohort4$Mono$rr
#> [1] 0.32
#> 
#> $Stage_Data$Cohort4$Mono$resp_bio
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  4  2  2  0  1  3  1
#> [26]  1  1  0  0  0  0
#> 
#> $Stage_Data$Cohort4$Mono$resp_hist
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  1  0  0  0  1  3  1
#> [26]  1  1  0  0  0  0
#> 
#> $Stage_Data$Cohort4$Mono$n
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  4  4  4  3  3  3  3
#> [26]  3  3  3  2  1  1
#> 
#> 
#> $Stage_Data$Cohort4$Back
#> $Stage_Data$Cohort4$Back$rr
#> [1] 0.32
#> 
#> $Stage_Data$Cohort4$Back$resp_bio
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  0  0  0  1  0  0  0
#> [26]  1  1  0  0  1  1
#> 
#> $Stage_Data$Cohort4$Back$resp_hist
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  0  0  0  1  0  1  0
#> [26]  1  1  0  0  1  0
#> 
#> $Stage_Data$Cohort4$Back$n
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  1  1  1  1  1  1  1
#> [26]  1  1  1  1  1  1
#> 
#> 
#> $Stage_Data$Cohort4$Plac
#> $Stage_Data$Cohort4$Plac$rr
#> [1] 0.12
#> 
#> $Stage_Data$Cohort4$Plac$resp_bio
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  0  1  1  1  0  0  0
#> [26]  0  0  0  0  0  1
#> 
#> $Stage_Data$Cohort4$Plac$resp_hist
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  0  0  0  0  0  0  0
#> [26]  0  0  0  0  0  0
#> 
#> $Stage_Data$Cohort4$Plac$n
#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA  1  1  1  1  1  1  1
#> [26]  1  1  1  1  1  1
#> 
#> 
#> $Stage_Data$Cohort4$prob_sup1
#> [1] 0.9533514 0.9945180 0.8919041 0.5000000 1.0000000 0.9999984 0.8199583
#> [8] 0.9923550
#> 
#> $Stage_Data$Cohort4$prob_sup2
#> [1] 0.9061979        NA 0.8168675 0.3652934 1.0000000        NA 0.6304993
#> [8] 0.9729672
#> 
#> $Stage_Data$Cohort4$prob_sup3
#> [1] 0.8299866        NA        NA        NA 1.0000000        NA        NA
#> [8]        NA
#> 
#> $Stage_Data$Cohort4$sup_interim_list
#> $Stage_Data$Cohort4$sup_interim_list[[1]]
#>      [,1] [,2]  [,3]  [,4]
#> [1,] TRUE TRUE FALSE FALSE
#> [2,] TRUE   NA  TRUE FALSE
#> [3,] TRUE   NA    NA    NA
#> 
#> 
#> $Stage_Data$Cohort4$fut_interim_list
#> list()
#> 
#> $Stage_Data$Cohort4$prom_interim_list
#> $Stage_Data$Cohort4$prom_interim_list[[1]]
#>       [,1]  [,2]  [,3]  [,4]
#> [1,] FALSE FALSE FALSE FALSE
#> [2,] FALSE    NA FALSE FALSE
#> [3,] FALSE    NA    NA    NA
#> 
#> 
#> $Stage_Data$Cohort4$sup_interim
#> [1] FALSE
#> 
#> $Stage_Data$Cohort4$fut_interim
#> [1] FALSE
#> 
#> $Stage_Data$Cohort4$prom_interim
#> [1] FALSE
#> 
#> $Stage_Data$Cohort4$interim_n
#> [1] 292
#> 
#> $Stage_Data$Cohort4$interim_n_cohort
#> [1] 54
#> 
#> $Stage_Data$Cohort4$sup_final_list
#> $Stage_Data$Cohort4$sup_final_list[[1]]
#>      [,1] [,2]  [,3] [,4]
#> [1,] TRUE TRUE FALSE TRUE
#> [2,] TRUE   NA FALSE TRUE
#> [3,] TRUE   NA    NA   NA
#> 
#> 
#> $Stage_Data$Cohort4$fut_final_list
#> list()
#> 
#> $Stage_Data$Cohort4$prom_final_list
#> $Stage_Data$Cohort4$prom_final_list[[1]]
#>       [,1]  [,2]  [,3]  [,4]
#> [1,] FALSE FALSE FALSE FALSE
#> [2,] FALSE    NA FALSE FALSE
#> [3,] FALSE    NA    NA    NA
#> 
#> 
#> $Stage_Data$Cohort4$sup_final
#> [1] FALSE
#> 
#> $Stage_Data$Cohort4$fut_final
#> [1] FALSE
#> 
#> $Stage_Data$Cohort4$prom_final
#> [1] FALSE
#> 
#> $Stage_Data$Cohort4$final_n
#> [1] 408
#> 
#> $Stage_Data$Cohort4$final_n_cohort
#> [1] 100
#> 
#> 
#>